Move tests to connect subdir.
