+++ test/connect/log.js
